# backend/src/routes/payments.js
// DCP Payment Routes — Moyasar SAR integration (DCP-31)
// Moyasar: Saudi-first gateway supporting mada, Apple Pay, VISA/MC in SAR
// Docs: https://moyasar.com/docs
const express = require('express');
const crypto = require('crypto');
const https = require('https');
const router = express.Router();
const db = require('../db');

function flattenRunParams(params) {
  if (params.length === 1 && Array.isArray(params[0])) return params[0];
  return params.reduce((acc, p) => (Array.isArray(p) ? acc.concat(p) : acc.concat([p])), []);
}

function runStatement(sql, ...params) {
  return db.prepare(sql).run(...flattenRunParams(params));
}

const MOYASAR_BASE = 'https://api.moyasar.com/v1';
const MOYASAR_SECRET = process.env.MOYASAR_SECRET_KEY || '';
const MOYASAR_WEBHOOK_SECRET = process.env.MOYASAR_WEBHOOK_SECRET || '';
const FRONTEND_URL = process.env.FRONTEND_URL || 'https://dcp.sa';

if (process.env.NODE_ENV === 'production' && !MOYASAR_WEBHOOK_SECRET) {
  throw new Error('MOYASAR_WEBHOOK_SECRET must be set in production');
}

function toFiniteNumber(value, { min = null, max = null } = {}) {
  const num = typeof value === 'number' ? value : Number(value);
  if (!Number.isFinite(num)) return null;
  if (min != null && num < min) return null;
  if (max != null && num > max) return null;
  return num;
}

function toFiniteInt(value, { min = null, max = null } = {}) {
  const num = toFiniteNumber(value, { min, max });
  if (num == null || !Number.isInteger(num)) return null;
  return num;
}

function normalizeCallbackUrl(value) {
  if (value == null || value === '') return null;
  if (typeof value !== 'string') return null;
  let parsed;
  try {
    parsed = new URL(value);
  } catch {
    return null;
  }
  if (!['https:', 'http:'].includes(parsed.protocol)) return null;
  return parsed.toString();
}

// ─── Moyasar API helper ────────────────────────────────────────────────────────

function moyasarRequest(method, path, body) {
  return new Promise((resolve, reject) => {
    if (!MOYASAR_SECRET) {
      return reject(new Error('MOYASAR_SECRET_KEY not configured'));
    }

    const bodyStr = body ? JSON.stringify(body) : null;
    const auth = Buffer.from(`${MOYASAR_SECRET}:`).toString('base64');
    const url = new URL(MOYASAR_BASE + path);

    const options = {
      hostname: url.hostname,
      path: url.pathname + url.search,
      method,
      headers: {
        'Authorization': `Basic ${auth}`,
        'Content-Type': 'application/json',
        'Accept': 'application/json',
      },
    };
    if (bodyStr) options.headers['Content-Length'] = Buffer.byteLength(bodyStr);

    const req = https.request(options, (res) => {
      let data = '';
      res.on('data', chunk => data += chunk);
      res.on('end', () => {
        try {
          const parsed = JSON.parse(data);
          if (res.statusCode >= 400) {
            const err = new Error(parsed.message || parsed.type || 'Moyasar API error');
            err.statusCode = res.statusCode;
            err.moyasarError = parsed;
            return reject(err);
          }
          resolve(parsed);
        } catch (e) {
          reject(new Error('Invalid Moyasar response: ' + data.slice(0, 200)));
        }
      });
    });

    req.on('error', reject);
    if (bodyStr) req.write(bodyStr);
    req.end();
  });
}

// ─── Webhook HMAC verification ─────────────────────────────────────────────────

function verifyMoyasarWebhook(rawBody, signatureHeader) {
  if (!MOYASAR_WEBHOOK_SECRET || !signatureHeader) return false;
  const signature = String(signatureHeader).trim().toLowerCase();
  const expected = crypto
    .createHmac('sha256', MOYASAR_WEBHOOK_SECRET)
    .update(rawBody)
    .digest('hex');
  try {
    return crypto.timingSafeEqual(
      Buffer.from(expected, 'hex'),
      Buffer.from(signature, 'hex')
    );
  } catch {
    return false;
  }
}

// ─── Renter auth helper ────────────────────────────────────────────────────────

function getRenter(req) {
  const key = req.headers['x-renter-key'] || req.query.key;
  if (!key) return null;
  return db.get('SELECT * FROM renters WHERE api_key = ? AND status = ?', key, 'active');
}

function requireRenter(req, res, next) {
  const renter = getRenter(req);
  if (!renter) {
    return res.status(401).json({ error: 'API key required (x-renter-key header or key query)' });
  }
  req.renter = renter;
  return next();
}

// ─── POST /api/payments/topup ──────────────────────────────────────────────────
// Initiate a SAR top-up via Moyasar. Returns a hosted checkout URL.
// Body: { amount_halala: number, payment_method: "creditcard"|"applepay" }
router.post('/topup', requireRenter, (req, res) => {
  const renter = req.renter;
  const { amount_halala, payment_method, amount_sar, source_type, callback_url } = req.body || {};
  const methodRaw = payment_method || source_type || 'creditcard';
  const paymentMethod = String(methodRaw).trim().toLowerCase();
  const allowedMethods = ['creditcard', 'applepay'];
  if (!allowedMethods.includes(paymentMethod)) {
    return res.status(400).json({ error: 'payment_method must be one of: creditcard, applepay' });
  }

  let amountHalala = toFiniteInt(amount_halala, { min: 100, max: 1000000 });
  if (amountHalala == null) {
    const legacyAmountSar = toFiniteNumber(amount_sar, { min: 1, max: 10000 });
    if (legacyAmountSar != null) {
      amountHalala = Math.round(legacyAmountSar * 100);
    }
  }
  if (amountHalala == null) {
    return res.status(400).json({ error: 'amount_halala must be an integer between 100 and 1000000' });
  }

  const amountSar = Number((amountHalala / 100).toFixed(2));
  const callbackUrl = normalizeCallbackUrl(callback_url) || `${FRONTEND_URL}/renter/billing?payment=callback`;
  if (callback_url != null && !normalizeCallbackUrl(callback_url)) {
    return res.status(400).json({ error: 'callback_url must be a valid http(s) URL' });
  }
  const description = `DCP balance top-up — ${renter.name} (${renter.email})`;

  const moyasarBody = {
    amount: amountHalala,
    currency: 'SAR',
    description,
    callback_url: callbackUrl,
    source: { type: paymentMethod },
    metadata: {
      renter_id: renter.id,
      renter_email: renter.email,
    },
  };

  moyasarRequest('POST', '/payments', moyasarBody)
    .then(payment => {
      const internalPaymentId = `pay_${crypto.randomBytes(12).toString('hex')}`;
      const moyasarId = payment.id;
      const paymentUrl = payment.source?.transaction_url || payment.source?.checkout_url || null;
      const now = new Date().toISOString();

      // Store payment record (status=pending until webhook confirms)
      runStatement(
        `INSERT INTO payments
           (payment_id, moyasar_id, renter_id, amount_sar, amount_halala, status, source_type, payment_method,
            description, callback_url, checkout_url, gateway_response, created_at)
         VALUES (?, ?, ?, ?, ?, 'pending', ?, ?, ?, ?, ?, ?, ?)`,
        internalPaymentId, moyasarId, renter.id, amountSar, amountHalala, paymentMethod, paymentMethod,
        description, callbackUrl, paymentUrl,
        JSON.stringify(payment), now
      );

      res.json({
        payment_url: paymentUrl,
        payment_id: internalPaymentId,
      });
    })
    .catch(err => {
      console.error('[payments] Moyasar topup error:', err.message, err.moyasarError);
      if (err.message === 'MOYASAR_SECRET_KEY not configured') {
        return res.status(503).json({
          error: 'Payment gateway not configured. Set MOYASAR_SECRET_KEY.',
          sandbox_hint: 'Use POST /api/payments/topup-sandbox for dev/test.',
        });
      }
      const statusCode = err.statusCode === 422 ? 422 : 502;
      res.status(statusCode).json({
        error: 'Payment initiation failed',
        details: err.moyasarError || err.message,
      });
    });
});

// ─── POST /api/payments/topup-sandbox ─────────────────────────────────────────
// Dev-only sandbox top-up: directly credits balance without Moyasar (when key not set).
// Disabled in production (requires MOYASAR_SECRET_KEY to be absent).
router.post('/topup-sandbox', (req, res) => {
  if (MOYASAR_SECRET) {
    return res.status(403).json({
      error: 'Sandbox top-up disabled when MOYASAR_SECRET_KEY is configured. Use /api/payments/topup.',
    });
  }

  const renter = getRenter(req);
  if (!renter) {
    return res.status(401).json({ error: 'API key required (x-renter-key header or key query)' });
  }

  const { amount_sar } = req.body;
  const amountSar = toFiniteNumber(amount_sar, { min: 0.01, max: 10000 });
  if (amountSar == null) {
    return res.status(400).json({ error: 'amount_sar must be between 0 and 10,000' });
  }

  const amountHalala = Math.round(amountSar * 100);
  const paymentId = 'sandbox-' + crypto.randomBytes(8).toString('hex');
  const now = new Date().toISOString();

  runStatement(
    `INSERT INTO payments
       (payment_id, moyasar_id, renter_id, amount_sar, amount_halala, status, source_type, payment_method,
        description, created_at, confirmed_at)
     VALUES (?, ?, ?, ?, ?, 'paid', 'sandbox', 'creditcard', ?, ?, ?)`,
    paymentId, paymentId, renter.id, amountSar, amountHalala,
    'Sandbox top-up (dev mode)', now, now
  );

  runStatement(
    `UPDATE renters SET balance_halala = balance_halala + ?, updated_at = ? WHERE id = ?`,
    amountHalala, now, renter.id
  );

  const updated = db.get('SELECT balance_halala FROM renters WHERE id = ?', renter.id);

  res.json({
    success: true,
    sandbox: true,
    payment_id: paymentId,
    amount_sar: amountSar,
    credited_halala: amountHalala,
    new_balance_sar: updated.balance_halala / 100,
    new_balance_halala: updated.balance_halala,
  });
});

// ─── POST /api/payments/webhook ────────────────────────────────────────────────
// Moyasar webhook handler. Verifies HMAC-SHA256 signature, credits balance on `paid`.
// Moyasar retries webhooks on non-2xx response.
// Raw body is mounted in server.js before express.json()
router.post('/webhook', (req, res) => {
  const rawBody = req.body; // Buffer (express.raw)
  const signature = req.headers['x-moyasar-signature'];

  if (!Buffer.isBuffer(rawBody)) {
    return res.status(500).json({ error: 'Webhook must be sent as raw application/json body' });
  }
  if (!MOYASAR_WEBHOOK_SECRET) {
    return res.status(503).json({ error: 'Webhook secret not configured' });
  }
  if (!verifyMoyasarWebhook(rawBody, signature)) {
    console.warn('[payments/webhook] Invalid HMAC signature — rejected');
    return res.status(401).json({ error: 'Invalid webhook signature' });
  }

  let event;
  try {
    event = JSON.parse(rawBody.toString('utf8'));
  } catch {
    return res.status(400).json({ error: 'Invalid JSON body' });
  }

  const paymentId = event.id;
  const status = event.status; // 'paid' | 'failed' | 'refunded' | 'initiated'
  const now = new Date().toISOString();

  // Look up stored payment record
  const payment = db.get(
    'SELECT * FROM payments WHERE moyasar_id = ? OR payment_id = ?',
    paymentId,
    paymentId
  );

  if (!payment) {
    // Unknown payment — return 200 to prevent Moyasar retries for stale events
    console.warn(`[payments/webhook] Unknown payment_id: ${paymentId} — acknowledging`);
    return res.json({ received: true, action: 'ignored_unknown' });
  }

  if (payment.status === status) {
    // Idempotent: already processed
    return res.json({ received: true, action: 'already_processed' });
  }

  if (status === 'paid') {
    // Credit renter balance once (idempotent by previous status guard)
    if (payment.status !== 'paid') {
      runStatement(
        `UPDATE renters SET balance_halala = balance_halala + ?, updated_at = ? WHERE id = ?`,
        payment.amount_halala, now, payment.renter_id
      );
    }
    runStatement(
      `UPDATE payments
       SET status = 'paid', confirmed_at = ?, gateway_response = ?
       WHERE payment_id = ?`,
      now, JSON.stringify(event), payment.payment_id
    );
    console.log(`[payments/webhook] Payment ${payment.payment_id} paid — credited ${payment.amount_halala} halala to renter ${payment.renter_id}`);
    return res.json({ received: true, action: 'balance_credited', amount_halala: payment.amount_halala });
  }

  if (status === 'failed') {
    runStatement(
      `UPDATE payments SET status = 'failed', gateway_response = ? WHERE payment_id = ?`,
      JSON.stringify(event), payment.payment_id
    );
    console.log(`[payments/webhook] Payment ${payment.payment_id} failed`);
    return res.json({ received: true, action: 'marked_failed' });
  }

  if (status === 'refunded') {
    const refundAmount = event.amount_refunded || payment.amount_halala;
    // Deduct from renter balance only if it was previously paid and not yet refunded
    if (payment.status === 'paid' && !payment.refunded_at) {
      const deduct = Math.min(refundAmount, payment.amount_halala);
      runStatement(
        `UPDATE renters SET balance_halala = MAX(0, balance_halala - ?), updated_at = ? WHERE id = ?`,
        deduct, now, payment.renter_id
      );
    }
    runStatement(
      `UPDATE payments SET status = 'refunded', refunded_at = ?, refund_amount_halala = ?, gateway_response = ? WHERE payment_id = ?`,
      now, refundAmount, JSON.stringify(event), payment.payment_id
    );
    console.log(`[payments/webhook] Payment ${payment.payment_id} refunded — ${refundAmount} halala`);
    return res.json({ received: true, action: 'refund_processed' });
  }

  // For any other status (e.g. 'initiated'), just update the gateway response
  runStatement(
    `UPDATE payments SET status = ?, gateway_response = ? WHERE payment_id = ?`,
    status, JSON.stringify(event), payment.payment_id
  );
  res.json({ received: true, action: 'status_updated', new_status: status });
});

// ─── GET /api/payments/verify/:paymentId ──────────────────────────────────────
// Fetch live payment status from Moyasar. Used by frontend to poll after redirect.
// Auth: renter key must own the payment.
router.get('/verify/:paymentId', (req, res) => {
  const renter = getRenter(req);
  if (!renter) {
    return res.status(401).json({ error: 'API key required' });
  }

  const { paymentId } = req.params;
  const localPayment = db.get(
    'SELECT * FROM payments WHERE renter_id = ? AND (payment_id = ? OR moyasar_id = ?)',
    renter.id, paymentId, paymentId
  );
  if (!localPayment) {
    return res.status(404).json({ error: 'Payment not found' });
  }

  // If already confirmed, return local record
  if (localPayment.status === 'paid') {
    return res.json({
      payment_id: localPayment.payment_id,
      status: 'paid',
      amount_sar: localPayment.amount_sar,
      amount_halala: localPayment.amount_halala,
      confirmed_at: localPayment.confirmed_at,
    });
  }

  // Fetch live status from Moyasar
  if (!MOYASAR_SECRET) {
    return res.json({
      payment_id: localPayment.payment_id,
      status: localPayment.status,
      amount_sar: localPayment.amount_sar,
      amount_halala: localPayment.amount_halala,
      note: 'Gateway not configured — showing local status',
    });
  }

  const externalPaymentId = localPayment.moyasar_id || localPayment.payment_id;
  moyasarRequest('GET', `/payments/${externalPaymentId}`, null)
    .then(payment => {
      const now = new Date().toISOString();

      // Sync local record if Moyasar reports paid
      if (payment.status === 'paid' && localPayment.status !== 'paid') {
        runStatement(
          `UPDATE renters SET balance_halala = balance_halala + ?, updated_at = ? WHERE id = ?`,
          localPayment.amount_halala, now, renter.id
        );
        runStatement(
          `UPDATE payments SET status = 'paid', confirmed_at = ?, gateway_response = ? WHERE payment_id = ?`,
          now, JSON.stringify(payment), localPayment.payment_id
        );
        console.log(`[payments/verify] Late sync: payment ${localPayment.payment_id} paid — credited ${localPayment.amount_halala} halala`);
      }

      res.json({
        payment_id: localPayment.payment_id,
        moyasar_id: payment.id,
        status: payment.status,
        amount_sar: payment.amount / 100,
        amount_halala: payment.amount,
        source_type: payment.source?.type,
        created_at: payment.created_at,
      });
    })
    .catch(err => {
      console.error('[payments/verify] Moyasar fetch error:', err.message);
      // Fallback to local record
      res.json({
        payment_id: localPayment.payment_id,
        moyasar_id: localPayment.moyasar_id || null,
        status: localPayment.status,
        amount_sar: localPayment.amount_sar,
        amount_halala: localPayment.amount_halala,
        gateway_error: 'Could not reach Moyasar — showing local status',
      });
    });
});

// ─── GET /api/payments/history ─────────────────────────────────────────────────
// Renter's own payment history (paginated).
router.get('/history', (req, res) => {
  const renter = getRenter(req);
  if (!renter) {
    return res.status(401).json({ error: 'API key required' });
  }

  const limit = Math.min(parseInt(req.query.limit) || 20, 100);
  const offset = Math.max(parseInt(req.query.offset) || 0, 0);

  const payments = db.all(
    `SELECT payment_id, moyasar_id, amount_sar, amount_halala, status, source_type, payment_method,
            description, created_at, confirmed_at, refunded_at, refund_amount_halala
     FROM payments WHERE renter_id = ?
     ORDER BY created_at DESC LIMIT ? OFFSET ?`,
    renter.id, limit, offset
  );

  const total = db.get('SELECT COUNT(*) as count FROM payments WHERE renter_id = ?', renter.id);
  const totalPaid = db.get(
    `SELECT COALESCE(SUM(amount_halala), 0) as total FROM payments WHERE renter_id = ? AND status = 'paid'`,
    renter.id
  );

  res.json({
    payments,
    pagination: { limit, offset, total: total.count },
    summary: {
      total_paid_sar: totalPaid.total / 100,
      total_paid_halala: totalPaid.total,
    },
  });
});

module.exports = router;