# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DCP Platform API
  version: 5.0.0
  description: |
    DCP is Saudi Arabia's first decentralized GPU compute marketplace.
    Providers register NVIDIA GPUs, install the Python daemon, and earn SAR.
    Renters submit compute jobs (LLM inference, image generation, training) that
    run on provider hardware. DCP takes 25% fee; providers earn 75%.

    ## Currency
    All monetary amounts are in **halala** (1 SAR = 100 halala) unless the field
    name ends in `_sar`.

    ## Auth
    Three auth schemes are used depending on the caller role:

    | Role | Scheme | How |
    |------|--------|-----|
    | Provider | `ProviderApiKey` | `?key=` query param or `x-provider-key` header |
    | Renter | `RenterApiKey` | `?key=` query param or `x-renter-key` header |
    | Admin | `AdminToken` | `x-admin-token` header |

    ## Rate Limits
    | Endpoint | Limit |
    |----------|-------|
    | `POST /providers/register` | 5 / IP / hour |
    | `POST /providers/heartbeat` | 4 / IP / minute |
    | `POST /renters/register` | 5 / IP / hour |
    | `POST /jobs/submit` | 10 / renter API key / minute |
    | `POST /payments/topup` | 10 / IP / minute |
    | `GET /api/providers/marketplace` | 30 / IP / minute |
    | `GET /api/admin/*` | 100 / IP / minute |
    | All other `/api/*` | 300 / IP / minute |

  contact:
    name: DCP DevRel
    url: https://dcp.sa/docs
  x-env-vars:
    DC1_ADMIN_TOKEN:
...
      required: true
    DC1_HMAC_SECRET:
...
      required: true
...
servers:
...
components:
  securitySchemes:
    ProviderApiKey:
      type: apiKey
      in: query
      name: key
      description: Provider API key. Prefix `dcp-provider-`. Also accepted as `x-provider-key` header.
    RenterApiKey:
      type: apiKey
      in: query
      name: key
      description: Renter API key. Prefix `dcp-renter-`. Also accepted as `x-renter-key` header.
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token set via DC1_ADMIN_TOKEN env var. Also accepted as Bearer token.
  schemas:
...
    Payment:
      type: object
      properties:
        payment_id:
          type: string
        amount_sar:
          type: number
        amount_halala:
          type: integer
        status:
          type: string
          enum: [initiated, paid, failed, refunded]
        source_type:
          type: string
          enum: [creditcard, mada, applepay, sandbox]
        description:
          type: string
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
paths:
...
  /api/payments/topup:
    post:
      tags: [Payments]
      summary: Initiate a Moyasar SAR payment
      description: |
        Creates a Moyasar payment and returns a hosted checkout URL. Renter must
        be redirected to `checkout_url` to complete payment. The webhook confirms
        and credits the balance. Rate limited to 10/min.
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_sar]
              properties:
                amount_sar:
                  type: number
                  minimum: 1
                  maximum: 10000
                  example: 100
                source_type:
                  type: string
                  enum: [creditcard, mada, applepay]
                  default: creditcard
                callback_url:
                  type: string
                  format: uri
                  nullable: true
                  description: Override default redirect URL after payment
      responses:
        '200':
          description: Payment initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  payment_id:
                    type: string
                  amount_sar:
                    type: number
                  amount_halala:
                    type: integer
                  status:
                    type: string
                  checkout_url:
                    type: string
                    format: uri
                  message:
                    type: string
        '503':
          description: Payment gateway not configured
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  sandbox_hint:
                    type: string
  /api/payments/topup-sandbox:
    post:
      tags: [Payments]
      summary: Dev-only sandbox balance top-up
      description: |
        Directly credits balance without Moyasar. Only available when
        `MOYASAR_SECRET_KEY` is **not** set (dev/staging environments). Disabled in production.
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_sar]
              properties:
                amount_sar:
                  type: number
                  minimum: 0.01
                  maximum: 10000
      responses:
        '200':
          description: Balance credited (sandbox)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sandbox:
                    type: boolean
                  payment_id:
                    type: string
                  amount_sar:
                    type: number
                  credited_halala:
                    type: integer
                  new_balance_sar:
                    type: number
                  new_balance_halala:
                    type: integer
        '403':
          description: Sandbox disabled in production
  /api/payments/webhook:
    post:
      tags: [Payments]
      summary: Moyasar payment webhook
      description: |
        Receives payment events from Moyasar. Verifies HMAC-SHA256 signature
        (`x-moyasar-signature` header). Credits balance on `paid` event.
        Must return 2xx or Moyasar will retry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Moyasar payment ID
                status:
                  type: string
                  enum: [initiated, paid, failed, refunded]
                amount:
                  type: integer
                  description: Amount in halala
                amount_refunded:
                  type: integer
                  nullable: true
      responses:
        '200':
          description: Webhook acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  action:
                    type: string
                    enum: [balance_credited, marked_failed, refund_processed, already_processed, ignored_unknown, status_updated]
        '401':
          description: Invalid HMAC signature
  /api/payments/verify/{paymentId}:
    get:
      tags: [Payments]
      summary: Verify payment status (post-redirect polling)
      description: Fetches live status from Moyasar. Used by frontend to poll after redirect back from checkout.
      security:
        - RenterApiKey: []
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: Payment not found or not owned by this renter
  /api/payments/history:
    get:
      tags: [Payments]
      summary: Renter payment history
      security:
        - RenterApiKey: []
      parameters:
        - in: query
          name: key
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated payment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                      offset:
                        type: integer
                      total:
                        type: integer
                  summary:
                    type: object
                    properties:
                      total_paid_sar:
                        type: number
                      total_paid_halala:
                        type: integer

  # ─── Admin ──────────────────────────────────────────────────────────────────
...